[refactor] Moved the MockAuthorisedFunctions to the mocks package
